pyproject.toml
[tool.black]
line-length = 79
target-version = ["py310"]

[tool.isort]
atomic = true
line_length = 79
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
balanced_wrapping = true
include_trailing_comma = true
multi_line_output = 3
known_third_party = [
    "pytest",
]
known_first_party = [
]

[tool.coverage.run]
branch = false
source = [
]
omit = [
]


[tool.mypy]
check_untyped_defs = true
#disallow_any_unimported = true
#disallow_any_expr = true
#disallow_any_decorated = true
disallow_any_generics = true
#disallow_any_explicit = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
#disallow_untyped_decorators = true
disallow_untyped_defs = true
follow_imports = "normal"
follow_imports_for_stubs = true
ignore_missing_imports = false
mypy_path = "./src"
namespace_packages = true
explicit_package_bases = true
show_column_numbers = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "conftest",
    "migrations.*",
    "tests.*",
    "*.tests",
]
ignore_errors = true

[tool.pytest.ini_options]
minversion = "6.0.0"
addopts = "-vv -s --tb=native"
testpaths = "tests/"


[tool.poetry]
name = "delta-lake-experiments"
version = "0.1.0"
description = ""
authors = ["zhukovgreen <iam@zhukovgreen.pro>"]
readme = "README.md"
packages = [{include = "delta_lake_experiments"}]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "3.2.1"
delta-spark = "2.0.0"
loguru = "^0.6.0"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
